Command now allocating memories for strings correctly. destroy_parse_res now freeing memory correctly and in correct order

+++ include/command.h
@@ -8,7 +8,7 @@ typedef struct Command {
     size_t argc; // number of args
 } Command;
 
-Command *create_command(int tokens_cnt, char **tokens);
+Command *create_command(size_t tokens_cnt, char **tokens);
 void destroy_command(Command *command);
 
 #endif
